feat(api): finish api for gift screens

+++ server/app/Http/Controllers/SuKienController.php
@@ -126,16 +126,26 @@ class SuKienController extends Controller
             else if ($request->type == 0)
             {
                 //Chi them 1 phan thuong
-                $phan_thuong = $request->phan_thuong;
+                $phan_thuongs = $request->phan_thuongs;
 
-                $phanThuong = PhanThuong::create([
-                    'type' => 0,
-                    'idSuKien' => $suKien->id,
-                ]);
+                if (count($phan_thuongs) != 1) {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'Sự kiện loại này được có 1 phần thưởng.'
+                    ], 403);
+                }
 
-                foreach($phan_thuong['items'] as $item)
+                foreach($phan_thuongs as $phan_thuong)
                 {
-                    $phanThuong->items()->attach($item['idItem'], ['soLuong' => $item['soLuong']]);
+                    $phanThuong = PhanThuong::create([
+                        'type' => 0,
+                        'idSuKien' => $suKien->id,
+                    ]);
+
+                    foreach($phan_thuong['items'] as $item)
+                    {
+                        $phanThuong->items()->attach($item['idItem'], ['soLuong' => $item['soLuong']]);
+                    }
                 }
             }
 
@@ -195,8 +205,6 @@ class SuKienController extends Controller
             $suKien->name = $request->name;
             $suKien->ngayBatDau = $request->ngayBatDau;
             $suKien->type = $request->type;
-            $suKien->save();
-
             
             if ($request->type == 1)
             {
@@ -208,7 +216,6 @@ class SuKienController extends Controller
                         $phanThuong->thanhTichHocTap = $phan_thuong['ThanhTichHocTap'];
                         $phanThuong->capHoc = $phan_thuong['capHoc'];
                         $phanThuong->type = 1;
-                        $phanThuong->save();
 
                         $phanThuong->items()->detach();
 
@@ -216,22 +223,33 @@ class SuKienController extends Controller
                         {
                             $phanThuong->items()->attach($item['idItem'], ['soLuong' => $item['soLuong']]);
                         }
+                        $phanThuong->save();
                     }
                 }
             } else if ($request->type == 0)
             {
-                $phan_thuong = $request->phan_thuong;
-                
-                $phanThuong = PhanThuong::find($phan_thuong['id']);
-                if ($phanThuong & $phanThuong->idSuKien == $idSuKien){
-                    $phanThuong->items()->detach();
-                    
-                    foreach($phan_thuong['items'] as $item)
-                    {
-                        $phanThuong->items()->attach($item['idItem'], ['soLuong' => $item['soLuong']]);
+                $phan_thuongs = $request->phan_thuongs;
+                if (count($phan_thuongs) != 1) {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'Sự kiện loại này được có 1 phần thưởng.'
+                    ], 403);
+                }
+                foreach($phan_thuongs as $phan_thuong)
+                {
+                    $phanThuong = PhanThuong::find($phan_thuong['id']);
+                    if ($phanThuong && $phanThuong->idSuKien == $idSuKien){
+                        $phanThuong->items()->detach();
+                        
+                        foreach($phan_thuong['items'] as $item)
+                        {
+                            $phanThuong->items()->attach($item['idItem'], ['soLuong' => $item['soLuong']]);
+                        }
                     }
                 }
             }
+
+            $suKien->save();
 
             return response()->json([
                 'data' => $suKien,
